+++ lib/pry/commands/change_inspector.rb
@@ -1,21 +1,21 @@
 class Pry::Command::ChangeInspector < Pry::ClassCommand
   match 'change-inspector'
-  group 'Input and Output'
-  description 'Change the current inspector proc.'
+  group 'Inspectors'
+  description 'Change the current inspector.'
   command_options argument_required: true
   banner <<-BANNER
     Usage: change-inspector NAME
 
-    Change the proc used to print return values. See list-inspectors for a list
-    of available procs and a short description of what each one does.
+    Change the current inspector. See 'list-inspectors' for a complete list of
+    available inspectors.
   BANNER
 
   def process(inspector)
     if inspector_map.key?(inspector)
       _pry_.print = inspector_map[inspector][:value]
-      output.puts "Switched to the '#{inspector}' inspector!"
+      output.puts "Switched to '#{inspector}' inspector."
     else
-      raise Pry::CommandError, "'#{inspector}' isn't a known inspector!"
+      raise Pry::CommandError, "'#{inspector}' isn't known."
     end
   end