+++ lib/pry/commands/change_prompt.rb
@@ -1,6 +1,6 @@
 class Pry::Command::ChangePrompt < Pry::ClassCommand
   match 'change-prompt'
-  group 'Input and Output'
+  group 'Prompts'
   description 'Change the current prompt.'
   command_options argument_required: true
   banner <<-BANNER
@@ -11,16 +11,29 @@ class Pry::Command::ChangePrompt < Pry::ClassCommand
   BANNER
 
   def process(prompt)
-    if prompt_map.key?(prompt)
-      _pry_.prompt = prompt_map[prompt][:value]
+    prompts = Pry::Prompt.all_by_name(prompt)
+    if prompts.size == 1
+      _pry_.prompt = prompts[0]
+    elsif prompts.size > 1
+      multiple_choice(prompts)
     else
       raise Pry::CommandError, "'#{prompt}' isn't a known prompt!"
     end
   end
 
-private
-  def prompt_map
-    Pry::Prompt::MAP
+  private
+  def multiple_choice(prompts)
+    _pry_.pager.page "Multiple aliases found, please choose:\n"
+    prompts.each.with_index(1) do |prompt, i|
+      _pry_.pager.page "#{i}) #{prompt.name} => #{prompt.alias_for}"
+    end
+    output.write "Choice: "
+    reply = (_pry_.input.respond_to?(:gets) ? _pry_.input.gets : _pry_.input.readline).strip
+    if reply =~ /^[1-9][0-9]*$/ and reply.to_i <= prompts.size
+      _pry_.prompt = prompts[reply.to_i-1]
+    else
+      multiple_choice(prompts)
+    end
   end
   Pry::Commands.add_command(self)
 end