+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ### HEAD
 
+* Add `Pry::Helpers::Text#{displayable_character?,displayable_string?,snowman_ready?}`.
+* `Pry::Helpers::Text#default` is deprecated in favor of `Pry::Helpers::Colors#strip_color`.
+* Coloring functions such as `black()` etc respect `_pry_.color` automatically, see [#1637](https://github.com/pry/pry/pull/1637).
+* Add `Pry#helpers`, see [#1635](https://github.com/pry/pry/pull/1635). Aliased as `Pry#h`.
+* Add `Pry::Helpers::Colors`, and a new `#paint` function.
+* Add clear-screen command ([#1630](https://github.com/pry/pry/pull/1630))
+* Fix [#1636](https://github.com/pry/pry/issues/1636)
+* Add 'alias-prompt' command.
+
+* Add `Pry::Prompt.[]`, `Pry::Prompt.add_prompt()`, `Pry::Prompt.alias_prompt()` and
+  `Pry::Prompt.remove_prompt()`, plus more, for integrating custom prompts with Pry
+  ([#1628](https://github.com/pry/pry/pull/1628)). See Pry::Prompt for complete details.
+
+* Add text helpers for background colors ([#1624](https://github.com/pry/pry/pull/1624))
+* Fix string literal methods completion. ([#1590](https://github.com/pry/pry/pull/1590))
+* Make sure Pry::WrappedModule::Candidate#source_location returns non-nil value when `.name` has
+  been redefined on a Class/Module ([#1623](https://github.com/pry/pry/pull/1623))
 * Add alias 'whereami[?!]+' for 'whereami' command. ([#1597](https://github.com/pry/pry/pull/1597))
 * Improve Ruby 2.4 support ([#1611](https://github.com/pry/pry/pull/1611)):
   * Deprecated constants are hidden from `ls` output by default, use the `-d` switch to see them.
@@ -28,7 +45,6 @@
 * Fixed error when Pry was trying to load incompatible versions of plugins ([#1312](https://github.com/pry/pry/issues/1312))
 * Fixed bug when `hist --clear` led to ArgumentError ([#1340](https://github.com/pry/pry/pull/1340))
 * Fixed the "uninitialized constant Pry::ObjectPath::StringScanner" exception during autocomplete ([#1330](https://github.com/pry/pry/issues/1330))
-* Secured usage of colours with special characters (RL_PROMPT_START_IGNORE and RL_PROMPT_END_IGNORE) in Pry::Helpers::Text ([#493](https://github.com/pry/pry/issues/493#issuecomment-39232771))
 * Fixed regression with `pry -e` when it messes the terminal ([#1387](https://github.com/pry/pry/issues/1387))
 * Fixed regression with space prefixes of expressions ([#1369](https://github.com/pry/pry/issues/1369))
 * Introduced the new way to define hooks for commands (with `Pry.hooks.add_hook("{before,after}_commandName")`). The old way is deprecated, but still supported (with `Pry.commands.{before,after}_command`) ([#651](https://github.com/pry/pry/issues/651))